+++ jest.config.ts
@@ -17,6 +17,7 @@ const config: Config.InitialOptions = {
     '\\.(css|less|scss|sass)$': 'identity-obj-proxy',
     '^@assets/(.*)$': '<rootDir>/src/assets/$1',
     '^@components/(.*)$': '<rootDir>/src/components/$1',
+    '^@decorators/(.*)$': '<rootDir>/src/decorators/$1',
     '^@lib/(.*)$': '<rootDir>/src/lib/$1',
     '^@services/(.*)$': '<rootDir>/src/services/$1',
     '^@styles/(.*)$': '<rootDir>/src/styles/$1',
