+++ tsconfig.json
@@ -22,6 +22,7 @@
       "@components/*": ["./src/components/*"],
       "@styles/*": ["./src/styles/*"],
       "@type/*": ["./src/types/*"],
+      "@decorators/*": ["./src/decorators/*"],
       "@lib/*": ["./src/lib/*"],
       "@services/*": ["./src/services/*"]
     }
